TODO CON BONUS

+++ libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: alegarci <alegarci@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/09 12:52:09 by alegarci          #+#    #+#             */
-/*   Updated: 2025/04/09 13:29:05 by alegarci         ###   ########.fr       */
+/*   Updated: 2025/04/15 18:05:48 by alegarci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *str1, const char *str2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n && (str1[i] || str2[i]))
